tools/requirements.txt
requests
python-dotenv
openai
pydantic
supabase
